Move Network to CorgiLib
